feat: neovim npm integration

+++ asap/archiso-asap/airootfs/usr/local/lib/asap/stages/stage2/2-09_copy-ssh-authorized-keys.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 # ----------------------------------------------------------------------------- #
 # Transfers the installation medium's authorized ssh keys to the new system's
-# asap resources directory, for later use in stage3.
+# asap resources directory, for later use in stage3, if they exist.
 # The keys will not be copied to '/mnt/root', as ssh access to the root user's
 # account is discouraged for security reasons. They will be stored in the asap
 # library folder and in stage3 installed to the non-root user, if so desired.
@@ -28,13 +28,16 @@ done
 
 ################################# BEGIN SCRIPT ##################################
 
-info_message "Storing authorized ssh keys in the asap library..."
+if [ -e /root/.ssh/authorized_keys ]; then
+    info_message "Storing authorized ssh keys in the asap library..."
 
-cp /root/.ssh/authorized_keys /mnt"$ASAP_RESOURCES_DIR"/authorized_keys.append
-if [ "$?" -ne 0 ]; then
-        exception_message "Failed to copy authorized_keys"
-        exit 1
+    cp /root/.ssh/authorized_keys /mnt"$ASAP_RESOURCES_DIR"/authorized_keys.append
+    if [ "$?" -ne 0 ]; then
+            exception_message "Failed to copy authorized_keys"
+            exit 1
+    fi
+
+    success_message "Stored the ssh keys temporarily."
 fi
 
-success_message "Stored the ssh keys temporarily."
 exit 0
